+++ packages/block-library/src/embed/index.js
@@ -8,15 +8,17 @@ import { getEmbedBlockSettings } from './settings';
 /**
  * WordPress dependencies
  */
-import { __ } from '@wordpress/i18n';
+import { __, _x } from '@wordpress/i18n';
 import { createBlock } from '@wordpress/blocks';
 
 export const name = 'core/embed';
 
 export const settings = getEmbedBlockSettings( {
-	title: __( 'Embed' ),
+	title: _x( 'Embed', 'block title' ),
 	description: __( 'The Embed block allows you to easily add videos, images, tweets, audio, and other content to your post or page.' ),
 	icon: embedContentIcon,
+	// Unknown embeds should not be responsive by default.
+	responsive: false,
 	transforms: {
 		from: [
 			{