+++ packages/editor/src/components/rich-text/index.native.js
@@ -20,11 +20,34 @@ import {
 	split,
 	toHTMLString,
 } from '@wordpress/rich-text';
+import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import { FORMATTING_CONTROLS } from './formatting-controls';
+
+const FORMATTING_CONTROLS = [
+	{
+		icon: 'editor-bold',
+		title: __( 'Bold' ),
+		format: 'bold',
+	},
+	{
+		icon: 'editor-italic',
+		title: __( 'Italic' ),
+		format: 'italic',
+	},
+	{
+		icon: 'admin-links',
+		title: __( 'Link' ),
+		format: 'link',
+	},
+	{
+		icon: 'editor-strikethrough',
+		title: __( 'Strikethrough' ),
+		format: 'strikethrough',
+	},
+];
 
 const isRichTextValueEmpty = ( value ) => {
 	return ! value || ! value.length;
@@ -42,24 +65,15 @@ export class RichText extends Component {
 		super( ...arguments );
 		this.onChange = this.onChange.bind( this );
 		this.onEnter = this.onEnter.bind( this );
+		this.onBackspace = this.onBackspace.bind( this );
 		this.onContentSizeChange = this.onContentSizeChange.bind( this );
 		this.changeFormats = this.changeFormats.bind( this );
 		this.toggleFormat = this.toggleFormat.bind( this );
 		this.onActiveFormatsChange = this.onActiveFormatsChange.bind( this );
-		this.onHTMLContentWithCursor = this.onHTMLContentWithCursor.bind( this );
 		this.state = {
 			formats: {},
 			selectedNodeId: 0,
 		};
-	}
-
-	// eslint-disable-next-line no-unused-vars
-	onHTMLContentWithCursor( htmlText, start, end ) {
-		if ( ! this.props.onSplit ) {
-			// TODO: insert the \n char instead?
-			return;
-		}
-		this.splitContent( htmlText, start, end );
 	}
 
 	/*
@@ -108,6 +122,12 @@ export class RichText extends Component {
 		if ( after ) {
 			after = this.valueToFormat( after );
 		}
+
+		// The onSplit event can cause a content update event for this block.  Such event should
+		// definitely be processed by our native components, since they have no knowledge of
+		// how the split works.  Setting lastEventCount to undefined forces the native component to
+		// always update when provided with new content.
+		this.lastEventCount = undefined;
 
 		onSplit( before, after );
 	}
@@ -177,8 +197,23 @@ export class RichText extends Component {
 		);
 	}
 
-	onEnter() {
-		this._editor.requestHTMLWithCursor();
+	// eslint-disable-next-line no-unused-vars
+	onEnter( event ) {
+		if ( ! this.props.onSplit ) {
+			// TODO: insert the \n char instead?
+			return;
+		}
+
+		this.splitContent( event.nativeEvent.text, event.nativeEvent.selectionStart, event.nativeEvent.selectionEnd );
+	}
+
+	// eslint-disable-next-line no-unused-vars
+	onBackspace( event ) {
+		if ( ! this.props.onBackspace ) {
+			return;
+		}
+
+		this.onBackspace( event.nativeEvent.text, event.nativeEvent.selectionStart, event.nativeEvent.selectionEnd );
 	}
 
 	shouldComponentUpdate( nextProps ) {
@@ -267,9 +302,9 @@ export class RichText extends Component {
 					text={ { text: html, eventCount: this.lastEventCount } }
 					onChange={ this.onChange }
 					onEnter={ this.onEnter }
+					onBackspace={ this.onBackspace }
 					onContentSizeChange={ this.onContentSizeChange }
 					onActiveFormatsChange={ this.onActiveFormatsChange }
-					onHTMLContentWithCursor={ this.onHTMLContentWithCursor }
 					color={ 'black' }
 					maxImagesWidth={ 200 }
 					style={ style }