+++ packages/editor/src/store/test/selectors.js
@@ -82,6 +82,7 @@ const {
 	isFirstMultiSelectedBlock,
 	getBlockMode,
 	isTyping,
+	isCaretWithinFormattedText,
 	getBlockInsertionPoint,
 	isBlockInsertionPointVisible,
 	isSavingPost,
@@ -89,7 +90,6 @@ const {
 	didPostSaveRequestFail,
 	getSuggestedPostFormat,
 	getEditedPostContent,
-	getNotices,
 	getReusableBlock,
 	isSavingReusableBlock,
 	isFetchingReusableBlock,
@@ -111,6 +111,7 @@ const {
 	INSERTER_UTILITY_HIGH,
 	INSERTER_UTILITY_MEDIUM,
 	INSERTER_UTILITY_LOW,
+	isPostSavingLocked,
 } = selectors;
 
 describe( 'selectors', () => {
@@ -888,6 +889,28 @@ describe( 'selectors', () => {
 			};
 
 			expect( isEditedPostPublishable( state ) ).toBe( true );
+		} );
+	} );
+
+	describe( 'isPostSavingLocked', () => {
+		it( 'should return true if the post has postSavingLocks', () => {
+			const state = {
+				postSavingLock: [ { 1: true } ],
+				currentPost: {},
+				saving: {},
+			};
+
+			expect( isPostSavingLocked( state ) ).toBe( true );
+		} );
+
+		it( 'should return false if the post has no postSavingLocks', () => {
+			const state = {
+				postSavingLock: [],
+				currentPost: {},
+				saving: {},
+			};
+
+			expect( isPostSavingLocked( state ) ).toBe( false );
 		} );
 	} );
 
@@ -2697,6 +2720,24 @@ describe( 'selectors', () => {
 		} );
 	} );
 
+	describe( 'isCaretWithinFormattedText', () => {
+		it( 'returns true if the isCaretWithinFormattedText state is also true', () => {
+			const state = {
+				isCaretWithinFormattedText: true,
+			};
+
+			expect( isCaretWithinFormattedText( state ) ).toBe( true );
+		} );
+
+		it( 'returns false if the isCaretWithinFormattedText state is also false', () => {
+			const state = {
+				isCaretWithinFormattedText: false,
+			};
+
+			expect( isCaretWithinFormattedText( state ) ).toBe( false );
+		} );
+	} );
+
 	describe( 'isSelectionEnabled', () => {
 		it( 'should return true if selection is enable', () => {
 			const state = {
@@ -3255,19 +3296,6 @@ describe( 'selectors', () => {
 			const content = getEditedPostContent( state );
 
 			expect( content ).toBe( '<!-- wp:default {\"modified\":true} /-->' );
-		} );
-	} );
-
-	describe( 'getNotices', () => {
-		it( 'should return the notices array', () => {
-			const state = {
-				notices: [
-					{ id: 'b', content: 'Post saved' },
-					{ id: 'a', content: 'Error saving' },
-				],
-			};
-
-			expect( getNotices( state ) ).toEqual( state.notices );
 		} );
 	} );