+++ packages/wordcount/src/index.js
@@ -92,8 +92,12 @@ function matchCharacters( text, regex, settings ) {
  */
 
 export function count( text, type, userSettings ) {
-	const settings = loadSettings( type, userSettings );
+	if ( '' === text ) {
+		return 0;
+	}
+
 	if ( text ) {
+		const settings = loadSettings( type, userSettings );
 		const matchRegExp = settings[ type + 'RegExp' ];
 		const results = ( 'words' === settings.type ) ?
 			matchWords( text, matchRegExp, settings ) :